1. Fixed error that was locking out the ability to use non-body boundary points. This was a result of an error-check gone too far!
2. Added hooks for some new thrusters. The actual thuster models are not contained in the open-source files.

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@300 77bb456d-b9d8-4424-8c4a-7c450a671a84

+++ trunk/src/mission.cpp
@@ -142,7 +142,7 @@ int mission::parse_outer_loop(int* Xouter)
 		vector<int> temp_sequence;
 		vector<int> temp_phase_type;
 
-		if (options.destination_list[j][0] > TheUniverse[j].bodies.size())
+		if (options.destination_list[j][0] > (int) TheUniverse[j].bodies.size())
 		{
 			std::cout << "ERROR: Journey " << j << " first body index " << options.destination_list[j][0] << " exceeds size of universe body list.Aborting." << std::endl;
 			throw 888;
@@ -218,7 +218,7 @@ int mission::parse_outer_loop(int* Xouter)
 		}
 
 		//encode the last phase of the journey
-		if (options.destination_list[j][1] > TheUniverse[j].bodies.size())
+		if (options.destination_list[j][1] > (int) TheUniverse[j].bodies.size())
 		{
 			std::cout << "ERROR: Journey " << j << " final body index " << options.destination_list[j][1] << " exceeds size of universe body list. Aborting." << std::endl;
 			throw 888;
